增加图片选择

+++ app/src/main/java/com/hlt/hao/grain/list/GrainListActivity.kt
@@ -11,6 +11,7 @@ import com.hlt.hao.BaseActivity
 import com.hlt.hao.R
 import com.hlt.hao.company.CompanyListActivity
 import com.hlt.hao.grain.add.AddGrainActivity
+import com.hlt.hao.grain.add.UploadImageActivity
 import com.hlt.hao.http.LiveDataResponse
 import kotlinx.android.synthetic.main.activity_grain_list.*
 
@@ -61,21 +62,45 @@ class GrainListActivity : BaseActivity() {
             getData()
 
         }
+        grainListAdapter.addChildClickViewIds(R.id.itemDelete, R.id.itemAddImage,R.id.viewImage)
 
         grainListAdapter.setOnItemChildClickListener { adapter, view, position ->
 
-            viewModel.deleteOrder(dataList[position].id.toString()).observe(this, Observer {
-                if (it.isState) {
-                    if (it.data.success) {
-                        Toast.makeText(applicationContext, "删除成功", Toast.LENGTH_SHORT).show()
-                        getData()
-                    } else {
-                        Toast.makeText(applicationContext, "删除失败", Toast.LENGTH_SHORT).show()
-
-                    }
+            if (view.id == R.id.itemAddImage) {
+                //上传图片
+                Intent(applicationContext, UploadImageActivity::class.java).let { intent->
+                    intent.putExtra("refId",dataList[position].id.toString())
+                    startActivity(intent)
                 }
+            } else if (view.id == R.id.itemDelete) {
+                //删除订单
+                viewModel.deleteOrder(dataList[position].id.toString()).observe(this, Observer {
+                    if (it.isState) {
+                        if (it.data.success) {
+                            Toast.makeText(applicationContext, "删除成功", Toast.LENGTH_SHORT).show()
+                            getData()
+                        } else {
+                            Toast.makeText(applicationContext, "删除失败", Toast.LENGTH_SHORT).show()
 
-            })
+                        }
+                    }
+
+                })
+            }else if (view.id==R.id.viewImage){
+                //查看图片
+                viewModel.getSingleImage(dataList[position].id.toString()).observe(this, Observer {
+                    if (it.isState) {
+                        if (it.data.success) {
+                            Toast.makeText(applicationContext, "获取", Toast.LENGTH_SHORT).show()
+                            getData()
+                        } else {
+                            Toast.makeText(applicationContext, "失败", Toast.LENGTH_SHORT).show()
+
+                        }
+                    }
+
+                })
+            }
         }
         add.setOnClickListener {
             startActivity(Intent(applicationContext, AddGrainActivity::class.java))
